%hhu isn't supported in ANSI C.

+++ rosettacode/100_doors.c
@@ -7,7 +7,7 @@ void print_doors(uint8_t doors[])
 {
     uint8_t door;
     for (door = 0; door < 100; door++)
-        printf("Door #%hhu: %s\n", door + 1, doors[door]? "open" : "closed");
+        printf("Door #%u: %s\n", (unsigned) door + 1, doors[door]? "open" : "closed");
 }
 
 void doors(void)
